services/vanna/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
groq==0.14.0
pydantic==2.5.3
pydantic-settings==2.1.0
